backend/requirements.txt
Django==4.2.13
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
python-decouple==3.8
celery==5.3.4
redis==5.0.1
drf-nested-routers==0.94.1
django-tenant-schemas==1.10.9
django-extensions==3.2.3
gunicorn==21.2.0
